refactor: break down json grammar parser in different files
